+++ src/components/Axis/index.js
@@ -1,3 +1,4 @@
 export { default as WXAxis } from './WXAxis/WXAxis.vue'
 export { default as WYAxis } from './WYAxis/WYAxis.vue'
 export { default as WZAxis } from './WZAxis/WZAxis.vue'
+export { default as WAxisLabel } from './WAxisLabel/WAxisLabel.vue'