+++ src/components/Common/index.js
@@ -1,2 +1,1 @@
 export { default as WDot } from './WDot/WDot.vue'
-export { default as WLabel } from './WLabel/WLabel.vue'