+++ src/register.js
@@ -14,7 +14,7 @@ import {
     WPie,
     WTooltip,
     WDot,
-    WLabel,
+    WAxisLabel,
     WBullet,
     WShowIfFit,
 } from './components'
@@ -47,7 +47,7 @@ export default (Vue, opts = {}) => {
         WPie,
         WTooltip,
         WDot,
-        WLabel,
+        WAxisLabel,
         WBullet,
         WShowIfFit,
     ]