+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.3.0] - 2019-09-10
 ### Added
 - WMarkers custom components
-- WLabel component
+- WAxisLabel component
 - 'visible' prop in WLine, WBar, WScatter components
 
 ### Changed